+++ src/Planners/src/ContactPhase.cpp
@@ -1,0 +1,17 @@
+/**
+ * @file ContactPhase.cpp
+ * @authors Stefano Dafarra
+ * @copyright 2020 Istituto Italiano di Tecnologia (IIT). This software may be modified and
+ * distributed under the terms of the GNU Lesser General Public License v2.1 or any later version.
+ */
+
+#include <BipedalLocomotion/Planners/ContactPhase.h>
+#include <algorithm>
+
+using namespace BipedalLocomotion::Planners;
+
+bool ContactPhase::isListIncluded(const std::string &key) const
+{
+    return  activeContacts.find(key) != activeContacts.end();
+}
+